feat: enhance mission management with creation functionality and refactor fetching logic
